% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{collapse_to_n_windows}
\alias{collapse_to_n_windows}
\title{Collapse Matrix To Windows by Window Number}
\usage{
collapse_to_n_windows(anno, mat, collapse_fun = mean, n_windows = 10)
}
\arguments{
\item{mat}{matrix}

\item{collapse_fun}{function for averaging values inside windows (Default = rowMeans, see Details).}

\item{n_windows}{number of windows (default=10, ie collapse to 10 values eah)}
}
\value{
matrix with n_windows columns
}
\description{
Averages Matrix Columns into Windows using an Averaging Function amd a defined number of windows
}
\details{
for a matrix with ncol=100 or 50 and n_windows=10, will return a matrix with ncol=10 in both cases. collapse_fun is a function that functions as ie rowMeans (the default) or rowSums.
}
\examples{

bedfile <- system.file("extdata", "snRNA_MS31_hg38.bed", package = "RMetaTools")
regions <- rtracklayer::import(bedfile)
mat <- get_single_strand_matrix(bw=bw, anno=regions, upstream=NULL, downstream=NULL)
dim(mat)
matc <- collapse_to_n_windows(anno, mat, mean, 10)
dim(matc)

}
