% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtracklayer.R
\name{create_bw_file_set}
\alias{create_bw_file_set}
\title{CreateBigWigSampleSet}
\usage{
create_bw_file_set(path, regex, junk_list, split_to, split_by)
}
\arguments{
\item{path}{path where the .bw files are located.}

\item{regex}{a regular expression (or list of) for selection files in path, default = '.bw'}

\item{junk_list}{a list of substrings to be removed before splitting to levels.}

\item{split_to}{split filenames into levels}

\item{split_by}{delimiter for splitting, default='_'}
}
\value{
A tidy data.frame with columns for path filename and all levels.
}
\description{
Creates a dataframe holding information on bigwig files.
}
\details{
Creates a data frame with path, filename. Uses the substrings in junk to clean-up the filename and the  simplified filename is used with split_to create levels using tidyr::separate function and split_by is the separator used. By default regex is always also removed from filename during simplification.
}
\examples{

path <- '/Users/schmidm/Documents/other_people_to_and_from/ClaudiaI/bw'
create_bw_file_set(path, c('_rep[3,4]', '.bw$'), c('_N20', '_3D12'), c('siRNA', 'ab', 'rep'), '_')

}
