% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMatrix.R
\name{get_single_strand_matrix}
\alias{get_single_strand_matrix}
\title{Single Strand Metagene Matrix}
\usage{
get_single_strand_matrix(bw, anno, upstream = 1000, downstream = 1000)
}
\arguments{
\item{bw}{filename for bigwig file.}

\item{anno}{GRanges object of regions to query.}

\item{upstream}{bp upstream of anchor (default=1000)}

\item{downstream}{bp downstream of anchor (default=1000)}
}
\value{
matrix with columns from most upstream to most downstream and rows are the individual regions from specified strand. rownames are set as name column from anno. colnames are set to relative position of each nt
}
\description{
Creates metagene matrix for single bigwig file.
}
\details{
Loads the bed file, extracts the position of the anchor point and creates ranges from x bp upstream to x bp downstream. I upstream and downstream are both NULL, collects signal within region specified in anno.
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000)
bw <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
mat <- get_single_strand_matrix(bw=bw, anno=regions, upstream=1000, downstream=5000)

}
