% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools_fun.R
\name{load_bwtool_agg}
\alias{load_bwtool_agg}
\title{load_bwtool_agg}
\usage{
load_bwtool_agg(path, regex, stranded = TRUE, junk_list = c(),
  split_to = c("siRNA", "ab", "rep"), split_by = "_", plus_cnt = 1,
  minus_cnt = 1)
}
\arguments{
\item{path}{path to aggregate files.}

\item{regex}{regular expression or list of regex, for selection of the files.}

\item{stranded}{search for 'plus' and 'minus' files on top of regex.}

\item{junk_list}{a list of substrings to be removed before splitting to levels.}

\item{split_to}{split filenames into levels}

\item{split_by}{delimiter for splitting, default='_'}

\item{plus_cnt}{count of features used for aggregate on the plus strand (default=1)}

\item{minus_cnt}{count of features used for aggregate on the minus strand (default=1)}
}
\value{
A tidy data frame.
}
\description{
Loads a bwtool aggregates file.
}
\details{
Loads a bwtool aggregate file and collects this in a \code{tidy} data.frame.
}
\examples{
 path <- '/Users/schmidm/Documents/other_people_to_and_from/ClaudiaI/bw/aggregates/exon1/'
 regex <- c('rep[3,4]', '_exon1_less200_', '.bed_starts_aggregate.txt$')
 junk_list <- c('_exon1_less200', '.bed_starts_aggregate.txt$', '_N20', '_3D12', '_plus', '_minus')
 df <- load_bwtool_agg(path, regex, stranded=TRUE, junk_list)
 ggplot(df, aes(x=rel_pos, y=value, color=siRNA, linetype=rep)) + geom_line() + facet_grid(.~ab)

 #for scaling:
 dplyr::group_by_(df, .dots=lapply(split_to, as.symbol)) \%>\%
 mutate(value = scale(value)) \%>\%
 ggplot(., aes(x=rel_pos, y=value, color=siRNA, linetype=rep)) + geom_line() + facet_grid(.~ab)

}
