% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools_fun.R
\name{load_bwtool_summary}
\alias{load_bwtool_summary}
\title{load_bwtool_summary}
\usage{
load_bwtool_summary(path, regex, header = TRUE, keep_bed = FALSE,
  with_quantiles = FALSE, with_sum = FALSE, with_sum_of_squares = FALSE,
  stranded = FALSE, junk_list = c(), split_to = c("siRNA", "ab", "rep"),
  split_by = "_")
}
\arguments{
\item{path}{path to summary files.}

\item{regex}{regular expression or list of regex, for selection of the files.}

\item{header}{logical. bwtool summary done with -header? Default=TRUE}

\item{keep_bed}{logical. bwtool summary done with -keep_bed? Default=FALSE}

\item{with_quantiles}{logical. bwtool summary done with -with-quantiles? Default=FALSE}

\item{with_sum}{logical. bwtool summary done with -with-sum? Default=FALSE}

\item{with_sum_of_squares}{logical. bwtool summary done with -with-sum-of-squares? Default=FALSE}

\item{stranded}{search for 'plus' and 'minus' files on top of regex.}

\item{junk_list}{a list of substrings to be removed before splitting to levels.}

\item{split_to}{split filenames into levels}

\item{split_by}{delimiter for splitting, default='_'}
}
\value{
A tidy data frame.
}
\description{
Loads a bwtool summary file.
}
\details{
Loads a bwtool summary file and collects this in a \code{tidy} data.frame.
}
\examples{
path <- '/Users/schmidm/Documents/other_people_to_and_from/ClaudiaI/bw/summaries/'
regex <- c('*rep[1,2]*_refseq_genes_summary.txt', 'FFL|Z18')
junk_list <- c('_refseq_genes_summary.txt', '_N20', '_3D12')
df <- load_bwtool_summary(path, regex, junk_list=junk_list, with_quantiles = T, with_sum = T, with_sum_of_squares = T)
ggplot(df, aes(x=rep, y=log2(sum), fill=rep)) + geom_violin() + facet_grid(siRNA~ab)


}
