% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptools_funs.R
\name{load_deeptoolsmatrix}
\alias{load_deeptoolsmatrix}
\title{load_deeptoolsmatrix}
\usage{
load_deeptoolsmatrix(fname, na.omit = TRUE, na.fill = 0,
  as.matrix = FALSE)
}
\arguments{
\item{na.omit}{Skip na values in the matrix.}

\item{na.fill}{Fills NAs with this value zero.}

\item{as.matrix}{Return matrix as matrix (default is FALSE, returns tidy data.frame)}

\item{file}{the file to load.}
}
\value{
An object of class \code{deeptools_matrix} with provided means and covariances
  computed as 1/\code{alpha}.
}
\description{
Loads a deeptools matrix file.
}
\details{
Loads a gzipped deeptools matrix file and collects this in a \code{tidy} data.frame. If as.matrix is true, the header information is added as attribute.
}
\examples{

 fname <- system.file("extdata", "matrix.gz", package = "RMetaTools")
 df <- load_deeptoolsmatrix(fname)
 head(df)
 df_withNA <- load_deeptoolsmatrix(fname, na.omit=FALSE)
 head(df_withNA)
 df_withNAas0 <- load_deeptoolsmatrix(fname, na.omit=FALSE, na.fill=0)
 head(df_withNAas0)

}
