% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptools_funs.R
\name{load_deeptoolsmatrix3}
\alias{load_deeptoolsmatrix3}
\title{load_deeptoolsmatrix3}
\usage{
load_deeptoolsmatrix3(fname, na.omit = TRUE, na.fill = 0,
  as.matrix = FALSE)
}
\arguments{
\item{fname}{the file to load.}

\item{na.omit}{Skip na values in the matrix. Default is TRUE.}

\item{na.fill}{Fills NAs with this value zero. Default is 0.}

\item{as.matrix}{Return matrix as matrix (default is FALSE, ie returns tidy data.frame instead)}
}
\value{
An object of class \code{deeptools_matrix} with provided means and covariances
  computed as 1/\code{alpha}.
}
\description{
Loads a deeptools v3 matrix file.
}
\details{
Loads a gzipped deeptools matrix file and collects this in a \code{tidy} data.frame. If as.matrix is true, the header information is added as attribute.
}
\examples{

 fname <- system.file("extdata", "deeptools3_matrix.gz", package = "RMetaTools")
 df <- load_deeptoolsmatrix3(fname)
 head(df)
 df_withNA <- load_deeptoolsmatrix3(fname, na.omit=FALSE)
 head(df_withNA)
 df_withNAas0 <- load_deeptoolsmatrix3(fname, na.omit=FALSE, na.fill=0)
 head(df_withNAas0)

}
