% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptools_funs.R
\name{load_deeptoolsmatrix3_new}
\alias{load_deeptoolsmatrix3_new}
\title{load_deeptoolsmatrix3_new}
\usage{
load_deeptoolsmatrix3_new(fname, na.fill)
}
\arguments{
\item{fname}{the file to load.}

\item{na.fill}{Fills NAs with this value zero. Default is NA.}
}
\value{
An object of class \code{metagene}.
}
\description{
Loads a deeptools v3 matrix file new version.
}
\details{
Always returns and S3 object. Loads a gzipped deeptools matrix file and collects this in a
  metagene object, with slots: $header, $regions, $samples,
  $groups and $matrices. Where $header contains the header information as a
  named vector. Matrices is a list of matrices where each matrix is specific
  to 1 sample and 1 group. $samples are the samplenames for each matrix.
  $groups are the groups for each matrix. $rel_pos is the list of relative positions to anchorpoint for each group. Names are the group names. $regions same as rel_pos are the region names (ie col4 of bed file).
}
\examples{

 fname <- system.file("extdata", "deeptools3_matrix.gz", package = "RMetaTools")
 m <- load_deeptoolsmatrix3_new(fname)
 m$header
 m$samples
 m$groups
 lapply(m$rel_pos, head)
 lapply(m$regions, head)
 lapply(m$matrices, dim)

}
