% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{mat_to_tbl}
\alias{mat_to_tbl}
\title{Tidy Metagene Matrix}
\usage{
mat_to_tbl(mat, anno, anchor, upstream, downstream, window_size)
}
\arguments{
\item{mat}{metagene matrix created from get_matrix}

\item{anno}{GRanges object of regions to query.}

\item{anchor}{'TSS', 'TES' or 'center.'}

\item{upstream}{bp upstream of anchor (default=1000)}

\item{downstream}{bp downstream of anchor (default=1000)}

\item{window_size}{size for binning in bp (default=1, ie no binning)}

\item{strand}{strand of anno to use (default='+')}
}
\value{
matrix with columns from most upstream to most downstream and rows are the individual regions.
}
\description{
Converts metagene matrix into a tidy tbl.
}
\details{
Loads the bed file, extracts the position of the anchor point and creates ranges from x bp upstream to x bp downstream.
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000, 50)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
mat <- get_matrix(bw_plus, bw_minus, regions, 1000, 1000, 50)
tidy_meta <- mat_to_tbl(mat)

}
