% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_match.R
\name{match_quantiles}
\alias{match_quantiles}
\title{Quantile Matching}
\usage{
match_quantiles(list_of_values, n_quantiles = 20)
}
\arguments{
\item{list_of_values}{list of sets of values.}

\item{n_quantiles}{number of quantiles used for matching (default=20 but check details)}
}
\value{
list with 2 elements: indices for each element in list_of_values
}
\description{
Selects from list of vectors, items from each vector with same count and quantile distribution
}
\details{
Splits the smallest set values into n_quantiles as defined. For each quantile gets a similar sized sample of values within this quantile from each element in list_of_values.
}
\examples{

values1 <- rnorm(10000)
values2 <- rnorm(100,2,1)
values3 <- rnorm(1000,3,2)
summary(values1)
summary(values2)
summary(values3)

matched_idx <- match_quantiles(list(values1, values2, values3))
summary(values1[matched_idx[[1]]])
summary(values2[matched_idx[[2]]])
summary(values3[matched_idx[[3]]])

}
