% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{meta_average}
\alias{meta_average}
\title{Create Metagene Averages}
\usage{
meta_average(tbl)
}
\arguments{
\item{tbl}{tidy metagene tbl}
}
\value{
tbl
}
\description{
Per position bin averages of specified tidy metagene tbl.
}
\details{
uses t.test to get mean ('estimate'), confidence interval etc for each position bin converted to columns using broom::tidy
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
anno=bedfile;anchor='center';upstream=1000;downstream=1000;window_size=10
collapse_fun = rowMeans;negate_neg_strand_values=FALSE
metamat <- metagene_matrix(bw_plus, bw_minus, bedfile, anchor, upstream, downstream, window_size)
tidy_avgs <- meta_average(metamat)

#alt:
tidy_log2_avgs <- meta_average(mutate(metamat, value = log2(value+1)))

}
