% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{meta_regions}
\alias{meta_regions}
\title{Metagene Regions}
\usage{
meta_regions(bed, anchor = "TSS", upstream = 1000, downstream = 1000)
}
\arguments{
\item{bed}{filename for bed file.}

\item{anchor}{anchor point TSS, TES, center or NULL (default='TSS')}

\item{upstream}{bp upstream of anchor (default=1000)}

\item{downstream}{bp downstream of anchor (default=1000)}
}
\value{
GRanges with regions.
}
\description{
Creates regions for metagene analysis.
}
\details{
Loads the bed file. If anchor is NULL, returns region from TSS-upstream to TES+downstream. If anchor, extracts the position of the anchor point and creates ranges from x bp upstream to x bp downstream of the anchor point.
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000)

}
