% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{metagene_aggregate}
\alias{metagene_aggregate}
\title{Metagene Aggregate}
\usage{
metagene_aggregate(m, aggregate_fun = "mean", transform_fun = NULL, ...)
}
\arguments{
\item{m}{the object to create the metagene aggregate for}

\item{aggregate_fun}{the function name to aggregate values such as 'events', 'sum', 'mean', 'ttest' (default is 'mean')}

\item{transform_fun}{a function to be applied to vlues before aggregate function is applied.}

\item{...}{other arguments passed to fun}
}
\value{
a tibble with column, rel_pos, sample, group, and aggregate values where column name is mean, sum, events or estimate+conf.low+conf.high for ttest
}
\description{
Metagene aggregate from an S3 metagene object.
}
\details{
Aggregates values.
transform_fun takes a matrix as input and returns a similar sized matrix. ie log2p1 <- function(m){log2(m+1)}
aggreagate_fun can be mean, sum, events or t.test or vector containing several of those.
}
\examples{
 fname <- system.file("extdata", "matrix_grouped.gz", package = "RMetaTools")
 m <- load_deeptoolsmatrix3_new(fname)
 metagene_aggregate(m)
 metagene_aggregate(m, aggregate_fun = 'events')
 metagene_aggregate(m, aggregate_fun = 'ttest')
 log2p1 <- function(m){log2(m+1)}
 metagene_aggregate(m, aggregate_fun = 'mean', transform_fun = log2p1)

}
