% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{metagene_ratio}
\alias{metagene_ratio}
\title{Metagene Ratio}
\usage{
metagene_ratio(m, samples, control, transform_fun = NULL, ...)
}
\arguments{
\item{m}{the object to create the metagene aggregate for}

\item{samples}{sample names for one or more matrices from m to be used}

\item{control}{sample name for one of the matrices from m to be used as denominator}

\item{transform_fun}{a function to be applied to vlues before aggregate function is applied.}

\item{...}{other arguments passed to fun}
}
\value{
a metagene object
}
\description{
Computes Ratios between matrices to one control
}
\details{
Ratio of values.
transform_fun takes a matrix as input and returns a similar sized matrix. ie p1 <- function(m){m+1}
Computes the ratio to the control matrix.
Handles all groups if there is more than one group present.
}
\examples{
 fname <- system.file("extdata", "matrix_grouped.gz", package = "RMetaTools")
 m <- load_deeptoolsmatrix3_new(fname)
 p1 <- function(m){return(m+1)}
 ratios <- metagene_ratio(m, c('RRP40', 'ZCCHC8', 'ZFC3H1'), 'MTR4', transform_fun=p1)
}
