% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{modify.metagene}
\alias{modify.metagene}
\title{Select and rename metagene object sample or group names}
\usage{
\method{modify}{metagene}(
  m,
  select_samples = NULL,
  remove_samples = NULL,
  select_groups = NULL,
  remove_groups = NULL,
  rename_samples = NULL,
  rename_groups = NULL,
  verbose = FALSE
)
}
\arguments{
\item{m}{a metagene object}

\item{select_samples}{regex or list of regex(es) for samples to select}

\item{remove_samples}{regex or list of regex(es) for samples to remove}

\item{select_groups}{regex or list of regex(es) for groups to remove}

\item{remove_groups}{regex or list of regex(es) for groups to remove}

\item{rename_samples}{a regexes and substitute string, default is substitute is empty}

\item{rename_groups}{a regexes and substitute string, default is substitute is empty}

\item{verbose}{report changes to stdout}
}
\value{
a tibble with columns, rel_pos, value, sample, group, and name, where name is the id column of the metagene object regions, usually col4 of the bed file used to create the object.
}
\description{
Select and/or rename sample and/or group names
}
\details{

}
\examples{
 fname <- system.file("extdata", "matrix_grouped.gz", package = "RMetaTools")
 m <- load_deeptoolsmatrix3_new(fname)

}
