% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptools_funs.R
\name{parse_matrix_header}
\alias{parse_matrix_header}
\title{parse_matrix_header}
\usage{
parse_matrix_header(fname)
}
\arguments{
\item{file}{the file to get the header from.}
}
\description{
Parses a deeptools matrix file header.
}
\details{
Loads a gzipped deeptools matrix file and collects this in a \code{tidy} data.frame.
}
