% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(df, sort_y = "sum", facets = NULL)
}
\arguments{
\item{df}{the df to load.}

\item{sort_y}{how to sort y axis of the heatmap ('region_len', 'sum', 'mean', 'max' )}

\item{facets}{a string for facetting of the plot.}
}
\value{
An ggplot2 plot object.
}
\description{
Plots tidy dataframe from a deeptools matrix file to a ggplot2 object.
}
\details{
Plots a df from a deeptools matrix tibble.
}
\examples{

 fname <- "/Users/schmidm/Documents/other_people_to_and_from/ClaudiaI/pA_seq/deeptools_out/Evgenia_PASseq_all_snRNA_joined.gz"
 df <- load_deeptoolsmatrix(fname)
 head(df)
 plot_heatmap(df, sort_y='sum', facets='.~sample_name')

 #can be further modified
 p <- plot_heatmap(df, sort_y='sum', facets='.~sample_name')
 p + ylab('')
 p + scale_fill_gradient2(low='white', high='darkred')
 p <- p + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())
 p + scale_fill_gradient2(low='white', high='darkred', name ='log10(norm. reads)')

}
