% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{plot_htmp}
\alias{plot_htmp}
\title{Plot Heatmap}
\usage{
plot_htmp(tbl, color_by = "value", do_interpolate = FALSE)
}
\arguments{
\item{tbl}{tidy metagene tbl}

\item{color_by}{name of column from tbl to use as value (default='value')}

\item{do_interpolate}{interpolate geom_raster output (default=FALSE)}
}
\value{
ggplot2 object
}
\description{
Heatmap of specified tidy metagene tbl.
}
\details{
uses ggplot2 to create plot of the heatmap
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
tidy_meta <- metagene_matrix(bw_plus, bw_minus, regions, 1000, 1000, 50)
plot_htmp(tidy_meta, color_by='value', do_interpolate=FALSE)
bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")

bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
anno=bedfile;anchor='center';upstream=1000;downstream=1000;window_size=10
collapse_fun = rowMeans;negate_neg_strand_values=FALSE
bedfile <- system.file("extdata", "snRNA_MS31_hg38.bed", package = "RMetaTools")
metamat <- metagene_matrix(bw_plus, bw_minus, bedfile, anchor, upstream, downstream, window_size)
plot_htmp(metamat, color_by='value', do_interpolate=FALSE)
metamat <- metagene_matrix(bw_plus, bw_minus, bedfile, anchor=NULL, 10, 10, collapse_to_n_wins=5, collapse_avg_fun=mean)
plot_htmp(metamat, color_by='value', do_interpolate=FALSE)

#log2 values incl pseudocount
metamat \%<>\% mutate(log2_value = log2(value+1))
plot_htmp(metamat, color_by='log2_value', do_interpolate=FALSE)

}
