% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot Metagene Average Profile}
\usage{
plot_profile(meta_tbl, color_by = NULL)
}
\arguments{
\item{color_by}{name of column from tbl to use for coloring of lines and ranges (default=NULL)}

\item{tbl}{tidy metagene tbl}
}
\value{
tbl
}
\description{
Plot Metagene Average Profile
}
\details{
plots line from ('estimate'), and range from conf.low and conf.high columns in tbl
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000, 50)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
mat <- get_matrix(bw_plus, bw_minus, regions, 1000, 1000, 50)
tidy_meta <- mat_to_tbl(mat)
tidy_avgs <- meta_average(tidy_meta)
plot_profile(tidy_avgs)

}
