% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMetaTools.R
\docType{package}
\name{RMetaTools}
\alias{RMetaTools}
\title{RMetaTools: Compendium of Tools for Metagene Analysis.}
\description{
Uses and depends on python-based deeptools and C-based bwtool.
}
\section{RMetaTools functions}{

to doo
}

