% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(df, value_name = "value", ctrl_factor, ctrl_name,
  pseudocount = NA, pseudocounting_method = "fill_with_min")
}
\arguments{
\item{df}{tidy data frame of values and conditions.}

\item{value_name}{column name of the column containing the values}

\item{ctrl_factor}{name of the factor containing the control}

\item{ctrl_name}{name of the control condition}

\item{pseudocount}{pseudocount value}

\item{pseudocount_method}{method used for pseudocounting (default is 'fill_with_min')}
}
\value{
A tidy data frame with all factors but containing sensitivity instead of values.
}
\description{
Computes Sensitivity for Metagene Plots.
}
\details{
Loads a \code{tidy} data.frame. Containing 1 column for each factor and all values in 1 column. Need to provide column name of the column containing the control (ie 'siRNA') and the name of the control (ie 'siFFL'). Computes the sensitivity: sensitivity = (expression(kd) + pseudocount) / (expression(ctrl) + pseudocount)$.
Pseudocounting: Either adds pseudocount to all values, or if pseudocount_method == "fill_with_min" uses the min of all values for pseudocounting.
}
\examples{
 path_to_RNAseq <- '/Users/schmidm/Documents/other_people_to_and_from/ClaudiaI/Effie_RNAseq_bws/RNAseq_deeptools_out/'
 m <- load_deeptoolsmatrix(paste0(path_to_RNAseq, 'snR_TES_joined.gz'), na.omit = FALSE, na.fill = 0) \%>\%  separate(sample_name, c('x', 'sname' ), sep='_') \%>\% mutate(sname = sub('rep-2plus', 'repplus', sname)) \%>\% separate(sname, c('siRNA', 'f', 'rep', 'bedstrand' ), sep='-') \%>\% dplyr::select(id, rel_pos, siRNA, value)
 head(m)
 sensitivity(m, ctrl_factor = 'siRNA', ctrl_value = 'siFFL',)

 #for scaling:
 dplyr::group_by_(df, .dots=lapply(split_to, as.symbol)) \%>\%
 mutate(value = scale(value)) \%>\%
 ggplot(., aes(x=rel_pos, y=value, color=siRNA, linetype=rep)) + geom_line() + facet_grid(.~ab)

}
