% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{to_tibble}
\alias{to_tibble}
\title{Coerce metagene object to tibble}
\usage{
to_tibble(m, ...)
}
\arguments{
\item{m}{a metagene object}

\item{...}{other arguments, currently ignored}
}
\value{
a tibble with columns, rel_pos, value, sample, group, and name, where name is the id column of the metagene object regions, usually col4 of the bed file used to create the object.
}
\description{
Collects info from metagene object into a single tibble.
}
\details{

}
\examples{
 fname <- system.file("extdata", "matrix_grouped.gz", package = "RMetaTools")
 m <- load_deeptoolsmatrix3_new(fname)
 to_tibble(m)
}
