% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{get_matrix}
\alias{get_matrix}
\title{Metagene Matrix}
\usage{
get_matrix(bw_plus, bw_minus, anno, upstream = 1000, downstream = 1000,
  window_size = 1)
}
\arguments{
\item{bw_plus}{filename for plus strand bigwig file.}

\item{bw_minus}{filename for minus strand bigwig file.}

\item{anno}{GRanges object of regions to query.}

\item{upstream}{bp upstream of anchor (default=1000)}

\item{downstream}{bp downstream of anchor (default=1000)}

\item{window_size}{size for binning in bp (default=1, ie no binning)}

\item{strand}{strand of anno to use (default='+')}
}
\value{
matrix with columns from most upstream to most downstream and rows are the individual regions. Rownames are the Name column from the bed file.
}
\description{
Creates metagene matrix for both strands using 1 or 2 bigwig files.
}
\details{
Loads the bed file, extracts the position of the anchor point and creates ranges from x bp upstream to x bp downstream.
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000, 50)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
mat <- get_matrix(bw_plus, bw_minus, regions, 1000, 1000, 50)

#if bigwigs are not stranded, ie from ChIP experiment both plus and minus are the same file
bw <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_hg38.bw", package = "RMetaTools")
mat <- get_matrix(bw, bw, regions, 1000, 1000, 50)

}
