% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{get_single_strand_matrix}
\alias{get_single_strand_matrix}
\title{Single Strand Metagene Matrix}
\usage{
get_single_strand_matrix(bw, anno, upstream = 1000, downstream = 1000,
  window = 1, strand = "+")
}
\arguments{
\item{bw}{filename for bigwig file.}

\item{anno}{GRanges object of regions to query.}

\item{upstream}{bp upstream of anchor (default=1000)}

\item{downstream}{bp downstream of anchor (default=1000)}

\item{strand}{strand of anno to use (default='+')}

\item{window_size}{size for binning in bp (default=1, ie no binning)}
}
\value{
matrix with columns from most upstream to most downstream and rows are the individual regions from specified strand.
}
\description{
Creates metagene matrix for single bigwig file.
}
\details{
Loads the bed file, extracts the position of the anchor point and creates ranges from x bp upstream to x bp downstream.
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000, 50)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
mat <- get_single_strand_matrix(bw_plus, regions, 1000, 1000, 50, '+')

}
