% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptools_funs.R
\name{load_deeptoolsmatrix3}
\alias{load_deeptoolsmatrix3}
\title{load_deeptoolsmatrix3}
\usage{
load_deeptoolsmatrix3(
  fname,
  na.omit = TRUE,
  na.fill = 0,
  as.matrix = FALSE,
  as.S3 = FALSE
)
}
\arguments{
\item{fname}{the file to load.}

\item{na.fill}{Fills NAs with this value zero. Default is 0.}

\item{as.matrix}{Return matrix as matrix (default is FALSE, ie returns tidy data.frame instead)}

\item{as.S3}{Return S3 class metagene object, see details for more info}
}
\value{
An object of class \code{deeptools_matrix} with provided means and covariances
  computed as 1/\code{alpha}.
}
\description{
Loads a deeptools v3 matrix file.
}
\details{
Loads a gzipped deeptools matrix file and collects this in a
  \code{tidy} data.frame. If as.matrix is true, the header information is
  added as attribute.
  as.S3 returns a metagene object, with slots: $header, $regions, $samples,
  $groups and $matrices. Where $header contains the header information as a
  named vector. Matrices is a list of matrices where each matrix is specific
  to 1 sample and 1 group. $samples are the samplenames for each matrix.
  $groups are the groups for each matrix.
}
\examples{

 fname <- system.file("extdata", "deeptools3_matrix.gz", package = "RMetaTools")
 df <- load_deeptoolsmatrix3(fname)
 head(df)
 df_withNAas0 <- load_deeptoolsmatrix3(fname, na.fill=0)
 head(df_withNAas0)
 s3obj <- load_deeptoolsmatrix3(fname, na.fill=0, as.S3=TRUE)
 head(s3obj$regions)
 lapply(s3obj$matrices, head)
 s3obj$samples
 s3obj$groups
 lapply(s3obj$rel_pos, head)

}
