% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{meta_average}
\alias{meta_average}
\title{Create Metagene Averages}
\usage{
meta_average(tbl)
}
\arguments{
\item{tbl}{tidy metagene tbl}
}
\value{
tbl
}
\description{
Per position bin averages of specified tidy metagene tbl.
}
\details{
uses t.test to get mean ('estimate'), confidence interval etc for each position bin converted to columns using broom::tidy
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
mat <- get_matrix(bw_plus, bw_minus, regions, 1000, 1000, 50)
tidy_meta <- mat_to_tbl(mat)
tidy_avgs <- meta_average(tidy_meta)

}
