% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{metagene_matrix}
\alias{metagene_matrix}
\title{Metagene Matrix}
\usage{
metagene_matrix(bw_plus, bw_minus, anno, anchor, upstream, downstream,
  window_size, negate_neg_strand_values = FALSE)
}
\arguments{
\item{bw_plus}{bigwig filename for plus strand signal}

\item{bw_minus}{bigwig filename for minus strand signal}

\item{anno}{Bed filename for annotations to use OR a GRanges object (see Details).}

\item{anchor}{'TSS', 'TES' or 'center.'}

\item{upstream}{bp upstream of anchor (default=1000)}

\item{downstream}{bp downstream of anchor (default=1000)}

\item{window_size}{size for binning in bp (default=1, ie no binning)}

\item{negate_neg_strand_values}{negate values from minus strand bigwig (default=FALSE).}

\item{strand}{strand of anno to use (default='+')}
}
\value{
Tidy data frame with columns: gene (name, ie col4 from *bed* file), rel_pos (position in bp relative to *anchor*), value (averaged value from bigwig file).
}
\description{
Computes a metagene matrix from scratch.
}
\details{
Loads the bed file, extracts the position of the anchor point and creates ranges from x bp upstream to x bp downstream. Queries from both bigwigs and the correct strand. Summarizes signal of window_size bins and returns a tidy data.frame. negate_neg_strand_values can be set to TRUE to deal with minus strand bigwigs which contain all negative values as used by some labs.
When using a GRanges object for anno, this assumes the bigwigs will be queries from start to end! All ranges must be exactly same size!
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
metamat <- metagene_matrix(bw_plus, bw_minus, bedfile, 'center', 1000, 1000, 50)

}
