% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{plot_htmp}
\alias{plot_htmp}
\title{Plot Heatmap}
\usage{
plot_htmp(tbl, color_by = "value", do_interpolate = FALSE)
}
\arguments{
\item{tbl}{tidy metagene tbl}

\item{color_by}{name of column from tbl to use as value (default='value')}

\item{do_interpolate}{interpolate geom_raster output (default=FALSE)}
}
\value{
ggplot2 object
}
\description{
Heatmap of specified tidy metagene tbl.
}
\details{
uses ggplot2 to create plot of the heatmap
}
\examples{

bedfile <- system.file("extdata", "Chen_PROMPT_TSSs_liftedTohg38.bed", package = "RMetaTools")
regions <- meta_regions(bedfile, 'TSS', 1000, 5000)
bw_plus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_plus_hg38.bw", package = "RMetaTools")
bw_minus <- system.file("extdata", "GSM1573841_mNET_8WG16_siLuc_minus_hg38.bw", package = "RMetaTools")
mat <- get_matrix(bw_plus, bw_minus, regions, 1000, 1000, 50)
tidy_meta <- mat_to_tbl(mat)
plot_htmp(tidy_meta, color_by='value', do_interpolate=FALSE)

#log2 values incl pseudocount
tidy_meta \%<>\% mutate(log2_value = log2(value+1))
plot_htmp(tidy_meta, color_by='log2_value', do_interpolate=FALSE)

}
